requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-crispy-forms==1.12.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-summernote==0.8.11.6
et-xmlfile==1.1.0
excel2json-3==0.1.6
idna==3.2
MarkupPy==1.14
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.1
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
